+++ custom_components/solarcharger/switch.py
@@ -73,21 +73,25 @@ class SolarChargerSwitchEntity(SolarChargerEntity, SwitchEntity, RestoreEntity):
     def turn_on(self, **kwargs: Any) -> None:
         """Turn the entity on."""
         self._attr_is_on = True
+        self.update_ha_state()
 
     # ----------------------------------------------------------------------------
     def turn_off(self, **kwargs: Any) -> None:
         """Turn the entity off."""
         self._attr_is_on = False
+        self.update_ha_state()
 
     # ----------------------------------------------------------------------------
     async def async_turn_on(self, **kwargs: Any) -> None:
         """Turn the entity on."""
-        self._attr_is_on = True
+        # self._attr_is_on = True
+        self.turn_on()
 
     # ----------------------------------------------------------------------------
     async def async_turn_off(self, **kwargs: Any) -> None:
         """Turn the entity off."""
-        self._attr_is_on = False
+        # self._attr_is_on = False
+        self.turn_off()
 
     # ----------------------------------------------------------------------------
     async def async_added_to_hass(self) -> None: